# docs/reference/openapi.kernel.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SoulOS Kernel",
    "version": "0.6.0"
  },
  "paths": {
...
    "/v1/avatars": {
      "post": {
        "tags": [
          "avatars"
        ],
        "summary": "Register Avatar",
        "operationId": "register_avatar_v1_avatars_post",
        "parameters": [
          {
            "name": "X-SoulOS-Account-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Soulos-Account-Id"
            }
          },
          {
            "name": "X-SoulOS-Gateway-Secret",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Soulos-Gateway-Secret"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/avatars/ensure": {
      "post": {
        "tags": [
          "avatars"
        ],
        "summary": "Ensure Avatar",
        "operationId": "ensure_avatar_v1_avatars_ensure_post",
        "parameters": [
          {
            "name": "X-SoulOS-Account-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Soulos-Account-Id"
            }
          },
          {
            "name": "X-SoulOS-Gateway-Secret",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Soulos-Gateway-Secret"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnsureAvatarRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/avatars/by-capability/{capability}": {
      "get": {
        "tags": [
          "avatars"
        ],
        "summary": "Avatars By Capability",
        "operationId": "avatars_by_capability_v1_avatars_by_capability__capability__get",
        "parameters": [
          {
            "name": "capability",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Capability"
            }
          },
          {
            "name": "X-SoulOS-Account-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Soulos-Account-Id"
            }
          },
          {
            "name": "X-SoulOS-Gateway-Secret",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Soulos-Gateway-Secret"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/soulpacks": {
      "get": {
        "tags": [
          "avatars"
        ],
        "summary": "Get Soulpacks",
        "operationId": "get_soulpacks_v1_soulpacks_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/avatars/import-soulpack": {
      "post": {
        "tags": [
          "avatars"
        ],
        "summary": "Import Soulpack Avatar",
        "operationId": "import_soulpack_avatar_v1_avatars_import_soulpack_post",
        "parameters": [
          {
            "name": "X-SoulOS-Account-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Soulos-Account-Id"
            }
          },
          {
            "name": "X-SoulOS-Gateway-Secret",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Soulos-Gateway-Secret"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportSoulPackRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EnsureAvatarRequest": {
        "properties": {
          "external_key": {
            "type": "string",
            "title": "External Key"
          },
          "soul": {
            "additionalProperties": true,
            "type": "object",
            "title": "Soul"
          },
          "runtime_config": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Runtime Config"
          }
        },
        "type": "object",
        "required": [
          "external_key",
          "soul"
        ],
        "title": "EnsureAvatarRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ImportSoulPackRequest": {
        "properties": {
          "pack_id": {
            "type": "string",
            "title": "Pack Id"
          },
          "external_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Key"
          },
          "msv_preset": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Msv Preset"
          },
          "persist": {
            "type": "boolean",
            "title": "Persist",
            "default": true
          },
          "runtime_config": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Runtime Config"
          }
        },
        "type": "object",
        "required": [
          "pack_id"
        ],
        "title": "ImportSoulPackRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}